xp int to long

+++ src/main/java/dev/abidux/survive/model/skills/skill/CappedSkill.java
@@ -25,7 +25,7 @@ public class CappedSkill extends Skill {
         ItemMeta meta = item.getItemMeta();
         meta.setDisplayName(name.legacyText);
         int level = skill.getLevel();
-        int xp = skill.getXp();
+        long xp = skill.getXp();
         if (level == levels.length - 1) {
             meta.setLore(Arrays.asList(
                     ChatColor.of("#fcba03") + "Nível: " + ChatColor.of("#a30808") + "M" + ChatColor.of("#bf0a0a") + "Á" + ChatColor.of("#eb1515") + "X"));
